fix: update ECS task def delete to set inactive
Update the ECS task definition delete workflow to set ACTIVE
tasks to INACTIVE and then to keep the 100 most recent
INACTIVE task definitions.

This is being done because of how Terraform plan operations
will fail if the Terraform state references an ECS task
definition that no longer exists.  With this approach, the
Terraform plan will no longer fail if the state references
an INACTIVE ECS task definition.

+++ bin/delete-ecs-task-defs.sh
@@ -20,10 +20,17 @@ set -euo pipefail
 # The sleep commands in the script are to avoid API throttling issues.
 #
 
-TASK_ARNS="$1"
+ACTION="$1"
+TASK_ARNS="$2"
 AWS_REGION="ca-central-1"
 CLUSTER_NAME="Forms"
 SERVICE_NAME="form-viewer"
+
+# Exit if the ACTION is invalid
+if [ "$ACTION" != "DEREGISTER" ] && [ "$ACTION" != "DELETE" ]; then
+    echo "❌ Invalid action: $ACTION"
+    exit 1
+fi
 
 # Get the currently active task definition ARN as a safety check
 ACTIVE_TASK_ARN="$(aws ecs describe-services \
@@ -42,21 +49,25 @@ do
         continue
     fi
 
-    echo  "🧹 Deleting: $TASK_ARN"
+    # To delete a task, it must be deregistered first
+    if [ "$ACTION" == "DEREGISTER" ] || [ "$ACTION" == "DELETE" ]; then
+        echo  "🗑️ Deregistering: $TASK_ARN"
+        aws ecs deregister-task-definition \
+            --task-definition "$TASK_ARN" \
+            --region "$AWS_REGION" \
+            --no-cli-pager > /dev/null 2>&1
+        sleep 2
+    fi
 
-    # Task definitions must be set to INACTIVE before they can be deleted
-    aws ecs deregister-task-definition \
-        --task-definition "$TASK_ARN" \
-        --region "$AWS_REGION" \
-        --no-cli-pager > /dev/null 2>&1
-    sleep 2
-
-    # Delete the INACTIVE task definition
-    aws ecs delete-task-definitions \
-        --task-definitions "$TASK_ARN" \
-        --region "$AWS_REGION" \
-        --no-cli-pager > /dev/null 2>&1
-    sleep 2
+    if [ "$ACTION" == "DELETE" ]; then
+        echo  "🧹 Deleting: $TASK_ARN"
+        aws ecs delete-task-definitions \
+            --task-definitions "$TASK_ARN" \
+            --region "$AWS_REGION" \
+            --no-cli-pager > /dev/null 2>&1
+        sleep 2
+    
+    fi
 done < "$TASK_ARNS"
 
 echo  "🎉 All done!"
